Fix warnings in xrGame

+++ src/xrGame/inventory_upgrade_property.cpp
@@ -35,7 +35,7 @@ void Property::construct( shared_str const& property_id, Manager& manager_r )
 	m_name = CStringTable().translate( pSettings->r_string( id(), "name" ) );
 	m_icon._set( pSettings->r_string(id(), "icon") );
 	Fvector3 color = READ_IF_EXISTS(pSettings, r_fvector3, id(), "color", Fvector3().set(255, 255, 255));
-	m_color = color_rgba(color.x, color.y, color.z, 255);
+	m_color = color_rgba((u32) color.x, (u32) color.y, (u32) color.z, 255);
 
 	// functor
 	LPCSTR functor_str = pSettings->r_string( id(), "functor" );
